backend/requirements.txt
python-dotenv
fastapi[all]
uvicorn
aiortc
pipecat-ai[google,silero]
opencv-python

